+++ packages/ui/src/components/tooltip/index.jsx
@@ -7,7 +7,6 @@ import {
   POINTERLEAVE,
   Dom,
   useMemo,
-  FOCUSIN,
   FOCUS,
   isString,
 } from "@elf-framework/sapa";