+++ packages/design-system/CHANGELOG.md
@@ -1,5 +1,14 @@
 # @elf-framework/design-system
 
+## 2.1.31
+
+### Patch Changes
+
+- add avatar component
+  update menu, tab, tabstrip component
+- Updated dependencies []:
+  - @elf-framework/design-tokens@2.1.31
+
 ## 2.1.29
 
 ### Patch Changes